+++ src/openapi/mod.rs
@@ -1,221 +1,12 @@
-use std::{
-    convert::Infallible, error::Error, fmt::Display, fs::File, io::BufReader, path::Path,
-    str::FromStr,
-};
+pub mod linting;
+pub mod operations;
+
+use std::{error::Error, fs::File, io::BufReader, path::Path};
 
 use indexmap::IndexMap;
-use nonempty::NonEmpty;
-use openapiv3::{OpenAPI, Operation, PathItem, ReferenceOr, RequestBody, Response, StatusCode};
+
+use openapiv3::{OpenAPI, ReferenceOr, RequestBody, Response, StatusCode};
 use serde_json::Value;
-
-use crate::{
-    examples::{need_example, ExamplePayloads},
-    Tag,
-};
-
-#[derive(PartialEq, Clone, Debug, Hash, Eq)]
-pub(crate) enum HttpMethod {
-    GET,
-    POST,
-    PUT,
-    DELETE,
-    PATCH,
-    HEAD,
-    OPTIONS,
-    TRACE,
-}
-
-impl Display for HttpMethod {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let s = format!("{:?}", self);
-        write!(f, "{}", s.to_lowercase())
-    }
-}
-
-impl HttpMethod {
-    fn all() -> Vec<HttpMethod> {
-        vec![
-            HttpMethod::GET,
-            HttpMethod::POST,
-            HttpMethod::PUT,
-            HttpMethod::DELETE,
-            HttpMethod::PATCH,
-            HttpMethod::HEAD,
-            HttpMethod::OPTIONS,
-            HttpMethod::TRACE,
-        ]
-    }
-}
-
-fn operation_for<'s>(
-    method: &HttpMethod,
-    path_item: &'s PathItem,
-) -> Option<&'s openapiv3::Operation> {
-    match method {
-        HttpMethod::GET => path_item.get.as_ref(),
-        HttpMethod::POST => path_item.post.as_ref(),
-        HttpMethod::PUT => path_item.put.as_ref(),
-        HttpMethod::DELETE => path_item.delete.as_ref(),
-        HttpMethod::HEAD => path_item.head.as_ref(),
-        HttpMethod::PATCH => path_item.patch.as_ref(),
-        HttpMethod::OPTIONS => path_item.options.as_ref(),
-        HttpMethod::TRACE => path_item.trace.as_ref(),
-    }
-}
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq)]
-
-pub(crate) struct OperationId(pub(crate) String);
-
-impl Display for OperationId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl FromStr for OperationId {
-    type Err = Infallible;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(OperationId(s.to_owned()))
-    }
-}
-
-impl TryFrom<&Operation> for OperationId {
-    type Error = ();
-
-    fn try_from(op: &Operation) -> Result<Self, Self::Error> {
-        op.operation_id
-            .as_ref()
-            .map(|id| OperationId(id.to_owned()))
-            .ok_or(())
-    }
-}
-
-fn capitalize(s: &str) -> String {
-    let mut done = false;
-    s.chars()
-        .map(|c| {
-            if done {
-                c
-            } else {
-                done = true;
-                c.to_ascii_uppercase()
-            }
-        })
-        .collect()
-}
-fn camelize(path: &str) -> String {
-    path.split_terminator('/').map(|s| capitalize(s)).collect()
-}
-
-impl OperationId {
-    fn for_method_and_path(method: &HttpMethod, path: &str) -> Self {
-        let id = format!("{}{}", method, camelize(path));
-        OperationId(id)
-    }
-}
-
-#[derive(Debug, Hash, Clone, PartialEq, Eq)]
-pub(crate) struct OperationContext {
-    pub id: OperationId,
-    pub method: HttpMethod,
-    pub path: String,
-    pub tags: Vec<Tag>,
-}
-
-#[derive(Debug)]
-pub(crate) struct OperationWithContext {
-    pub context: OperationContext,
-    pub operation: Operation,
-}
-
-struct OperationIterator {
-    path_items: Vec<(String, PathItem)>,
-    current_item_operations: Option<NonEmpty<OperationWithContext>>,
-    current_item: usize,
-}
-
-impl OperationIterator {
-    fn new(path_items: Vec<(String, PathItem)>) -> Self {
-        OperationIterator {
-            path_items,
-            current_item_operations: None,
-            current_item: 0,
-        }
-    }
-}
-
-impl Iterator for OperationIterator {
-    type Item = OperationWithContext;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        //we have iterated over all the path items, nothing else to do
-        if self.current_item > self.path_items.len() {
-            None
-        } else {
-            //if there's some operations for the current PathItem,
-            if let Some(current_item_ops) = self.current_item_operations.take() {
-                //set current item operations to the operations' tail
-                self.current_item_operations = NonEmpty::from_vec(current_item_ops.tail);
-
-                Some(current_item_ops.head)
-            } else {
-                let mut next_op = None;
-
-                loop {
-                    if &self.current_item >= &self.path_items.len() {
-                        break;
-                    }
-
-                    let (path, path_item) = &self.path_items[self.current_item];
-                    self.current_item += 1;
-
-                    let operations: Option<NonEmpty<OperationWithContext>> = NonEmpty::from_vec(
-                        operations_for(&path, &path_item)
-                            .into_iter()
-                            .map(|(context, operation)| OperationWithContext {
-                                context,
-                                operation: operation.to_owned(),
-                            })
-                            .collect(),
-                    );
-
-                    if let Some(operations) = operations {
-                        self.current_item_operations = NonEmpty::from_vec(operations.tail);
-                        next_op = Some(operations.head);
-                        break;
-                    }
-                }
-
-                next_op
-            }
-        }
-    }
-}
-
-fn operations_for<'s>(
-    path: &str,
-    path_item: &'s openapiv3::PathItem,
-) -> Vec<(OperationContext, &'s openapiv3::Operation)> {
-    HttpMethod::all()
-        .iter()
-        .filter_map(|method| {
-            operation_for(method, path_item).map(|operation| {
-                let id = OperationId::try_from(operation)
-                    .unwrap_or_else(|_| OperationId::for_method_and_path(method, path));
-                let tags = operation.tags.iter().map(|s| Tag(s.to_owned())).collect();
-                let context = OperationContext {
-                    id,
-                    tags,
-                    method: method.to_owned(),
-                    path: path.to_owned(),
-                };
-                (context, operation)
-            })
-        })
-        .collect()
-}
 
 pub(crate) fn spec_from_file<P: AsRef<Path>>(path: P) -> Result<OpenAPI, Box<dyn Error>> {
     let file = File::open(path.as_ref())?;
@@ -277,156 +68,14 @@ pub(crate) fn clone_items<T: Clone>(
     values_by_key
 }
 
-#[derive(Debug, Default)]
-pub(crate) struct Stats {
-    total_operations: u16,
-    matching_filters: u16,
-    operations_needing_examples: u16,
-    operations_with_invalid_examples: u16,
-}
-
-impl Display for Stats {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "API Stats\n")?;
-        writeln!(
-            f,
-            "{:<15}: {:>3}",
-            "Total operations", self.total_operations
-        )?;
-        writeln!(
-            f,
-            "{:<15}: {:>3}",
-            "Matching filters", self.matching_filters
-        )?;
-        writeln!(
-            f,
-            "{:<15}: {:>3}",
-            "Needing examples", self.operations_needing_examples
-        )?;
-        writeln!(
-            f,
-            "{:<15}: {:>3}",
-            "Invalid examples", self.operations_with_invalid_examples
-        )?;
-
-        Ok(())
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct LintingIssues {
-    //TODO: replace box in favour of an enum
-    pub issues: NonEmpty<Box<dyn Error>>,
-}
-
-type OperationLintingIssues = IndexMap<OperationContext, LintingIssues>;
-
-#[derive(Debug)]
-pub(crate) enum LintingOutcome {
-    AllGood(Stats),
-    IssuesFound {
-        stats: Stats,
-        operation_linting_issues: OperationLintingIssues,
-    },
-    OperationNotFound(OperationId),
-}
-
-pub(crate) async fn lint(
-    spec: &OpenAPI,
-    tags: Vec<Tag>,
-    operation_id: Option<OperationId>,
-) -> LintingOutcome {
-    let components = spec
-        .components
-        .as_ref()
-        .map(Components::new)
-        .unwrap_or_default();
-
-    let mut stats = Stats::default();
-    let mut operation_linting_issues = IndexMap::new();
-
-    let path_items: Vec<(String, PathItem)> = clone_items(&spec.paths).into_iter().collect();
-
-    let cli_filter: Box<dyn Fn(&Operation) -> bool> = match (&operation_id, &tags) {
-        (Some(operation_id), _) => Box::new(|operation: &Operation| {
-            operation.operation_id == Some(operation_id.to_string())
-        }),
-        (_, tags) => Box::new(|operation: &Operation| {
-            if tags.len() == 0 {
-                true
-            } else {
-                tags.iter().any(|tag| operation.tags.contains(&tag.0))
-            }
-        }),
-    };
-
-    let operations: Vec<OperationWithContext> = OperationIterator::new(path_items).collect();
-    stats.total_operations = operations.len() as u16;
-
-    let operations: Vec<OperationWithContext> = operations
-        .into_iter()
-        .filter(|op| cli_filter(&op.operation))
-        .collect();
-
-    stats.matching_filters = operations.len() as u16;
-
-    //shortcut with OperationNotFound if supplied filter yields no result
-    if let (Some(operation_id), 0) = (operation_id.clone(), operations.len()) {
-        return LintingOutcome::OperationNotFound(operation_id);
-    }
-
-    let (operation_examples_result, needing_example_result) = futures::join!(
-        ExamplePayloads::from_operations(&operations, &components),
-        need_example(&operations)
-    );
-
-    if let Err(operation_report) = operation_examples_result {
-        let operation_report: OperationLintingIssues = operation_report
-            .into_iter()
-            .map(|(op_id, errors)| {
-                let issues: NonEmpty<Box<dyn Error>> = errors.map(|e| e.into());
-
-                let issues = LintingIssues { issues };
-                (op_id.to_owned(), issues)
-            })
-            .collect();
-
-        stats.operations_with_invalid_examples += operation_report.len() as u16;
-
-        operation_linting_issues.extend(operation_report);
-    }
-
-    if let Err(needing_examples) = needing_example_result {
-        stats.operations_needing_examples += needing_examples.len() as u16;
-
-        for (operation_id, schema_needing_example) in needing_examples {
-            if let Some(LintingIssues { issues }) = operation_linting_issues.get_mut(&operation_id)
-            {
-                issues.push(schema_needing_example.into());
-            } else {
-                let linting_issues = LintingIssues {
-                    issues: NonEmpty::new(schema_needing_example.into()),
-                };
-                operation_linting_issues.insert(operation_id, linting_issues);
-            }
-        }
-    }
-
-    if operation_linting_issues.len() == 0 {
-        LintingOutcome::AllGood(stats)
-    } else {
-        LintingOutcome::IssuesFound {
-            stats,
-            operation_linting_issues,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::io::Write;
 
-    use super::*;
+    use super::linting::*;
+    use super::operations::*;
+    use crate::openapi::*;
+    use openapiv3::PathItem;
     use tempdir::TempDir;
 
     #[test]
@@ -526,7 +175,6 @@ mod tests {
 
         assert_eq!(operation_ids, vec!["getA", "postA", "deleteA", "getB"])
     }
-
     #[test]
     fn lint_fails_on_unknown_operation_id() {
         let operation_id = OperationId("unknown".to_owned());
